Add files via upload
This data is all about Loan which is given on an online Platform and it is Peer to Peer Lending. So, there may be Chances of defaulting. So, we need to get the insights of the data of Which grades are defaulting and whether we should provide them loan or not and if we are providing loan then at what interest rates we should provide.